Feedback on the cleaner script
Your script works great, but I modified the input section to
be more in keeping with how Linux tools normally work. Generally
if you need input like a directory name, string to modify, etc
you would just pass it in on the commandline instead of specifically
asking the user for it. You want to prompt the user in situations
that would require extensive user input. For instance, if they were
filling out a form, putting their life's information on the
command line would be impractical. However, for most of our
purposes, passing in an argument will be more than sufficient.

+++ finished/clean_zero_files_python.py
@@ -1,25 +1,39 @@
 #!/usr/bin/env python
 
-#Input the path you wish to clean
-#This will recursively find all files
+# Input the path you wish to clean
+# This will recursively find all files
 # of 0 size and remove them
 
 import sys
 import os
 
-#main method
+# main method
 def main():
+    """
+    Michael cleans things up
+    """
 
-    #gets path from user
-    path = str(input("Enter the path surrounded by quotes (\"p/a/t/h...\"): "))
+    # gets path passed in from the user
+    args = sys.argv
+    if len(args) != 2:
+        # if we got here then they either entered too many args or not enough
+        print("You must supply one directory to clean")
 
-    #loops through the directory and removes any files of size 0
-    for root, dirs, files in os.walk(path):
-        for name in files:
-            filename = os.path.join(root,name)
-            if os.stat(filename).st_size == 0:
-                print("removing ", filename)
-                os.remove(filename)
+        # time to bail
+        sys.exit()
+    else:
+        # ok, looking good. let's remove some files
+        path = sys.argv[1]
+        print("Cleaning Directory : %s" % path)
+
+        # loops through the directory and removes any files of size 0
+        for root, dirs, files in os.walk(path):
+            for name in files:
+                filename = os.path.join(root, name)
+                if os.stat(filename).st_size == 0:
+                    print("  removing :", filename)
+                    os.remove(filename)
+
 
 if __name__ == "__main__":
     main()
